modified:   my_video.py 	modified:   server.py

+++ server.py
@@ -14,7 +14,7 @@ import time
 import os
 import subprocess
 import datetime
-from my_video import video_convert, video_merge, video_send_to_telegram, video_delete_files, delete_first
+from my_video import video_convert, video_merge, video_send_to_telegram, video_delete_files, delete_first, video_delete_all
 
 PORT = '8823'
 
@@ -183,13 +183,19 @@ async def call_charge_mode(request):
 	
 async def call_video(request):
 	script_path = '/home/pi/telegram_rover/capture/'
-	delete_first(script_path,10)
-	h264_files,mp4_files = video_convert(script_path)
-	if len(mp4_files):
-		video_merge(script_path, mp4_files)
+	if os.path.exists(script_path+'mp4/out.mp4'):
 		video_send_to_telegram(script_path)
-		video_delete_files(script_path,h264_files, mp4_files)
-	return web.Response(text=str(len(mp4_files))+' merged',content_type="text/html")
+		video_delete_all(script_path)
+		result = 'video file sent'
+	else:
+		delete_first(script_path,10)
+		h264_files,mp4_files = video_convert(script_path)
+		if len(mp4_files):
+			video_merge(script_path, mp4_files)
+			video_send_to_telegram(script_path)
+			video_delete_files(script_path,h264_files, mp4_files)
+			result = str(len(mp4_files))+' merged'
+	return web.Response(text=result,content_type="text/html")
 
 async def call_check(request):
 	return web.Response(text='ok',content_type="text/html")
